feat(gateway-protocol): add new deploy rpc
The deploy process rpc was able to deploy non-process resources (like
dmn) simply by passing in a different resource. Even though that works,
the related metadata cannot be responded, because it is process
specific (i.e. decision metadata cannot be provided). A new deploy rpc
is therefore necessary.

The deploy process rpc is replaced, and therefore deprecated. It will
still work as before, but we recommend all client maintainers to switch
to using the new deploy rpc.

The new deploy rpc is designed to be expandable to provide the metadata
of new types in the future. This could be achieved using the oneof
keyword.