refactor(protocol): rename Record#clone to Record#copyOf
Renames the `clone` method to `copyOf`. This was flagged as warning
as the clone method is a special method in Java, with specific semantics,
and the `Cloneable` interface is simply a marker interface to reuse the
built-in cloning facitilities. As we were always implementing the method
and changing the semantics, it's more accurate to use a copy method
instead.