feat(gateway): support deploy rpc
A new rpc can be independently implemented by the clients, but that's
not true for the gateway. It should immediately support the new rpc.

Implementation was pretty straightfoward, with lots of code duplicated
from the existing DeployProcess rpc. The only difference happens in the
ResponseMapper, where it also maps decisions and required decisions.