feat(msgpack-value): introduce stream on array value
It would be useful to have a standard way to transform the iterable
value array into a java stream. Obviously this then brings some useful
features like map, filter, etc.

We should always watch out with creating this Stream, because it is not
a thread-safe iterable. This makes it al the more important that we have
a standard way of creating a stream out of it, that makes sure its
avoiding Stream.parallel.